Bug 11752: display the correct frequency for serial subscriptions in OPAC details
This fixes bootstrap and prog by modifying the description displayed
in the OPAC's detail of serials.

RM NOTE: this patch does not cover the case where custom serial
frequencies have been defined.

TESTING to reproduce
- create/find a serial with a 1/week periodicity (4 in the database)
- Find it in the opac-detail.pl, click "more details" at the bottom
- validate the string.  Before the patch, it will say:
"The current subscription began on 2013-12-06 and is issued every 3
 weeks for 26 issues"

The "every 3 weeks" is clearly wrong.
In fact any periodicity chosen would display a wrong description, not
matching the staff interface.

After the patch, the display is corrected.

As a bonus, the "every 2 years" now has a description, where it had
none before.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>