Bug 1047: Punctuation in site names no longer breaks Edit/Delete z39.50 server administration
The template used to show the Z39.50 server administration page had a
bug that caused it to not correctly escape generated query strings.
Because the Z39.50 server name is used to lookup the server in order
to edit or delete it, when the server name is not passed correctly in
the query string, it is impossible to bring up the edit or delete forms
(without manually entering the escaped string.)

This patch corrects which template is filter used to escape those query
strings.

To test:
(1) Login to intranet, go to Administration -> Z39.50 servers
(2) Select "New Z3.50 Server". Enter a server name that contains an
    ampersand (&), e.g.: "FOO & BAR". Enter other details and submit.
    Click OK to confirmation message.
(3) In the server list, click on the server name, the "Edit" or "Delete"
    buttons for the server.

The correct and full server details should be brought up.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>