Bug 11935: fix capitalization: OPAC prog theme, checkouts tab
opac-user.tt in the prog theme contains 'Renew Selected' and 'Renew
All'. These are incorrectly capitalized. This patch corrects.

To test you must have OpacRenewalAllowed enabled. Apply the patch,
switch to the prog theme, and log in to the OPAC as a patron who has
checkouts.

At the bottom of the table of checkouts you should see the renew buttons
with correct capitalization.

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
String change, all good.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>