Bug 11957: reomve disused sample_only_param_tables.sql
This sql file tries to insert data into nonexistent tables.
It is useless and can be removed.

Test plan:
  git grep sample_only_param_tables
should not return any result.

NOTE: This is basically a git rm. Simple patch.

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
File is not used by the installers and it wouldn't work
with today's Koha for trying to insert data into non-existant
tables and tables with different structure.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>